Use usePostJustPublished instead of registry subscribe
